문제 너무 많아서 수정
