--- bustimeline/env
@@ -1,1 +1,0 @@
-SERVICE_KEY=NtJsRbGCOEUzaVWZV0YH%2BohG0HsEnbQsxLeHFL3y%2F1HHMeyRALLGqSGZs3qMMIFG%2FBoyQ8KBgfs29NRRsTwsOw%3D%3D
